Failed to finish feature
multiple supporting peices are in place now. I will need to refactor
certain things with the ast/exprs, try to specify the exact type and
provide alternate Enums for if only two possible like function args self
vs symbol.

This will also help with the difference between when symbols are being
initalized, referenced in an Info, or accessed/assigned